+++ modules/log/index.js
@@ -9,7 +9,7 @@ glog.create = function (name) {
 		// index 0 is always "Error"
 		for (var i = 1, len = stack.length; i < len; i++) {
 			if (stack[i].indexOf('gracelog/index.js') === -1 && stack[i].indexOf('gracenode/modules/log/index.js') === -1) {
-				name = stack[i].substring(stack[i].lastIndexOf('(') + 1, stack[i].lastIndexOf('.'));
+				name = stack[i].substring(stack[i].indexOf('/'), stack[i].lastIndexOf('.'));
 				break;
 			}
 		}