+++ core/index.js
@@ -16,6 +16,7 @@ var logCleaners = [];
 
 module.exports.Gracenode = Gracenode;
 
+
 function Gracenode() {
 	EventEmitter.call(this);
 	// listeners