+++ CHANGELOG.md
@@ -10,7 +10,7 @@ None
 
 #### log module's default log name does not have the full file path
 
-#### Updated dependency gracelog version from 0.1.10 to 0.1.12
+#### Updated dependency gracelog version from 0.1.10 to 0.1.13
 
 #### Minor bug fix in daemon log