+++ CHANGELOG.md
@@ -8,6 +8,8 @@ None
 
 ## Changed
 
+#### script lint.sh refactored
+
 #### log module's default log name does not have the full file path
 
 #### Updated dependency gracelog version from 0.1.10 to 0.1.13
@@ -18,7 +20,7 @@ None
 
 None
 
- ## Removed
+## Removed
 
 None